Align conversion test suite with eco-gotests conventions
Extract shared helpers (readResource, assertCRDServesVersion, nestedNumber)
into a dedicated helpers.go file, unify the namespace parameter pattern
across bootstrap and controlplane resource builders, and match the Labels
slice construction to the convention used by sibling suites.
