+++ tests/assisted/ztp/capoa-conversion/internal/tsparams/capoa-conversion-vars.go
@@ -8,7 +8,7 @@ import (
 
 var (
 	// Labels represents the range of labels that can be used for test cases selection.
-	Labels = append(append([]string{}, ztpparams.Labels...), LabelSuite)
+	Labels = append(ztpparams.Labels, LabelSuite)
 
 	// ReporterNamespacesToDump tells to the reporter from where to collect logs.
 	ReporterNamespacesToDump = map[string]string{
